clean up example files
